fix: remoting encoding and decoding error.

+++ naive-rpc-core/src/main/java/com/joizhang/naiverpc/remoting/transport/RequestHandlerRegistry.java
@@ -32,7 +32,7 @@ public class RequestHandlerRegistry {
         }
     }
 
-    public RequestHandler get(int type) {
+    public RequestHandler get(byte type) {
         return handlerMap.get(type);
     }
